+++ aws:/v0.19-file/test/wal/version
@@ -1,0 +1,1 @@
+v0.19-file
