+++ src/partykit/gateway.ts
@@ -1,7 +1,7 @@
 import PartySocket, { PartySocketOptions } from "partysocket";
 import { Result, URI, BuildURI, KeyedResolvOnce, runtimeFn, exception2Result } from "@adviser/cement";
 import { bs, ensureLogger, getStore, Logger, rt, SuperThis } from "@fireproof/core";
-URI.protocolHasHostpart("partykit:");
+
 export class PartyKitGateway implements bs.Gateway {
   readonly logger: Logger;
   readonly sthis: SuperThis;
@@ -261,6 +261,7 @@ export class PartyKitTestStore implements bs.TestGateway {
 }
 
 export function registerPartyKitStoreProtocol(protocol = "partykit:", overrideBaseURL?: string) {
+  URI.protocolHasHostpart(protocol);
   return bs.registerStoreProtocol({
     protocol,
     overrideBaseURL,
