+++ vitest.aws.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from "vitest/config";
+import tsconfigPaths from "vite-tsconfig-paths";
+
+export default defineConfig({
+  plugins: [tsconfigPaths()],
+  test: {
+    name: "aws",
+    exclude: [
+      "node_modules/@fireproof/core/tests/react/**",
+      "node_modules/@fireproof/core/tests/fireproof/config.test.ts",
+    ],
+    include: ["node_modules/@fireproof/core/tests/**/*test.?(c|m)[jt]s?(x)"],
+    globals: true,
+    setupFiles: "./setup.aws.ts",
+  },
+});
