+++ vitest.workspace.ts
@@ -4,6 +4,7 @@ import betterSqlite3 from "./vitest.better-sqlite3.config.ts";
 import nodeSqlite3Wasm from "./vitest.node-sqlite3-wasm.config.ts";
 import s3 from "./vitest.s3.config.ts";
 import partykit from "./vitest.partykit.config.ts";
+import aws from "./vitest.aws.config.ts";
 import connector from "./vitest.connector.config.ts";
 // import cf_kv from "./vitest.cf-kv.config.ts";
 
@@ -12,6 +13,7 @@ export default defineWorkspace([
   // betterSqlite3,
   // connector,
   //s3,
-  partykit,
+  aws,
+  // partykit,
   //cf_kv
 ]);
